Move `lemma_finite_int_set_has` into `vstd_extra`
